Update resource based on PR feedback
